Optimize scanner (#241)
Rather than hash the entire file on track index, only hash the first 1KB.
Queue a job to fully hash the entire file later. If the full hash is needed
in a location, which is currently only during track creation for deduplication,
calculate it on demand there if the hash of the first 1KB matches.

This speeds up indexing dramatically. Indexing used to take ~0.1sec per track
on a slow HDD. Now it is many times faster, and it is feasible to index large
libraries on the scale of minutes.